# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Ticket Place App",
        "contact": {
            "email": "ticket.place@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apid.c4m.mg/ticket-place-app/public",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/event-type-place": {
            "get": {
                "tags": [
                    "TypePlaces"
                ],
                "summary": "Get All Type place of events",
                "description": "return list of the all Type place of events",
                "operationId": "c7d7bba829a0ab513ac5e97d0ecb60ea",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action unauthorized"
                    },
                    "404": {
                        "description": "aucun résultat trouvé"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
        "/api/event-type-place/{type_place_id}": {
            "get": {
                "tags": [
                    "TypePlaces"
                ],
                "summary": "Show the type place ",
                "description": "Show type place and his event",
                "operationId": "8ec9ac599efadaefb0224ff9e624dcb6",
                "parameters": [
                    {
                        "name": "type_place_id",
                        "in": "path",
                        "description": "ID du types de place",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "opération succès"
                    },
                    "404": {
                        "description": "Type Place non trouvé"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "TypePlaces"
                ],
                "summary": "Update type place ",
                "description": "Update type place info of a event",
                "operationId": "08893c5c18a192fe61135a9cebfc115b",
                "parameters": [
                    {
                        "name": "type_event_id",
                        "in": "path",
                        "description": "ID du types de place",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "description": "Le nom du type de place",
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "description": "Le nombre de places pour ce type",
                                        "type": "integer"
                                    },
                                    "prix": {
                                        "description": "Le prix de ce type de place",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Types de place ajoutés avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Événement non trouvé"
                    }
                }
            }
        },
        "/api/event-type-place/{type_event_id}": {
            "delete": {
                "tags": [
                    "TypePlaces"
                ],
                "summary": "Delete type place event",
                "description": "Delete type place event",
                "operationId": "ff5d324fe65349731aa49d8c1bd7f790",
                "parameters": [
                    {
                        "name": "type_event_id",
                        "in": "path",
                        "description": "ID of the event to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "action not authorized"
                    },
                    "403": {
                        "description": "action forbiden"
                    },
                    "500": {
                        "description": "erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}